+++ packages/performer-playground/src/hooks/use-performer-client.ts
@@ -6,11 +6,17 @@ import {
   PerformerDeltaEvent,
   PerformerEvent,
   PerformerMessageEvent,
+  PerformerOptions,
 } from "@performer/core";
 import { jsx } from "@performer/core/jsx-runtime";
 import consola from "consola";
 
-export function usePerformerClient(app: Component<any> | null) {
+const logger = consola.withTag("usePerformerClient");
+
+export function usePerformerClient(
+  app: Component<any> | null,
+  options: PerformerOptions = {},
+) {
   const [events, setEvents] = useState<PerformerEvent[]>([]);
   const [performer, setPerformer] = useState<Performer | null>(null);
 
@@ -18,7 +24,7 @@ export function usePerformerClient(app: Component<any> | null) {
     if (!performer) {
       return;
     }
-    performer.input({ role: "user", content: [{ type: "text", text }] });
+    performer.input({ role: "user", content: text });
   }
 
   useEffect(() => {
@@ -27,15 +33,12 @@ export function usePerformerClient(app: Component<any> | null) {
     }
     try {
       setEvents([]);
-      const performer = new Performer(jsx(app, {}));
+      const performer = new Performer(jsx(app, {}), options);
       setPerformer(performer);
 
       performer.addEventListener("*", (event) => {
-        consola.info(
-          `usePerformerClient event=${event.type} message="Received event"`,
-        );
         if (event instanceof PerformerMessageEvent) {
-          return setEvents((prevEvents) => {
+          setEvents((prevEvents) => {
             const previous = prevEvents.findLast(
               (prev): prev is PerformerDeltaEvent =>
                 prev instanceof PerformerDeltaEvent &&
@@ -43,46 +46,52 @@ export function usePerformerClient(app: Component<any> | null) {
             );
 
             if (previous) {
-              consola.info(
-                `usePerformerClient event=${event.type} message="Found delta, discarding message." role=${event.detail.message.role} uid=${event.detail.uid}`,
+              logger.debug(
+                `event=${event.type} message="Found delta, discarding message." uid=${event.detail.uid} detail=${JSON.stringify(event.detail)}`,
               );
               return prevEvents;
             } else {
-              consola.info(
-                `usePerformerClient event=${event.type} message="Adding message." content="${event.detail.message.content}" role=${event.detail.message.role} uid=${event.detail.uid}`,
+              logger.debug(
+                `event=${event.type} message="Adding message." uid=${event.detail.uid} detail=${JSON.stringify(event.detail)}`,
               );
               return [...prevEvents, event];
             }
           });
         } else if (!(event instanceof PerformerDeltaEvent)) {
-          consola.info(`usePerformerClient event=${event.type}`);
-          return setEvents((prevEvents) => [...prevEvents, event]);
+          logger.debug(`event=${event.type}`);
+          setEvents((prevEvents) => [...prevEvents, event]);
         } else {
-          setEvents((prevEvents) => {
-            const previous = prevEvents.findLast(
-              (prev): prev is PerformerDeltaEvent =>
-                prev instanceof PerformerDeltaEvent &&
-                prev.detail.uid === event.detail.uid,
-            );
-            if (!previous) {
-              consola.info(
-                `usePerformerClient event=${event.type} message="No previous delta, adding new." uid=${event.detail.uid}`,
-              );
-              return [...prevEvents, event];
-            }
-            consola.info(
-              `usePerformerClient event=${event.type} message="Previous delta found, concatenating" uid=${event.detail.uid}`,
-            );
-            concatDelta(previous.detail.delta, event.detail.delta);
-            return prevEvents.toSpliced(prevEvents.length - 1, 1, previous);
-          });
+          setEvents((prevEvents) => applyDeltaEvent(prevEvents, event));
         }
       });
       performer.start();
     } catch (e) {
-      console.error("Performer exception", e);
+      logger.error("Performer exception", e);
     }
   }, [app]);
 
   return { events, sendMessage };
+}
+
+function applyDeltaEvent(
+  prevEvents: PerformerEvent[],
+  event: PerformerDeltaEvent,
+) {
+  const previousIndex = prevEvents.findIndex(
+    (prev): prev is PerformerDeltaEvent =>
+      prev instanceof PerformerDeltaEvent &&
+      prev.detail.uid === event.detail.uid,
+  );
+  const previous = previousIndex > -1 && prevEvents[previousIndex];
+  if (!previous) {
+    logger.debug(
+      `event=${event.type} message="No previous delta, adding new." uid=${event.detail.uid} detail=${JSON.stringify(event.detail)}`,
+    );
+    return [...prevEvents, event];
+  }
+  logger.debug(
+    `event=${event.type} message="Previous delta found, concatenating" uid=${event.detail.uid} detail=${JSON.stringify(event.detail)}`,
+  );
+  concatDelta(previous.detail.delta, event.detail.delta);
+  return prevEvents.toSpliced(previousIndex, 1, previous);
 }